Change sched_scan logic for large number of configured networks (UPSTREAM)
One of the major motivations behind sched scan is offloading
periodic scanning to the device firmware so that the host can
suspend and only be woken up when a known network is close by.
When there are more than max_match_sets or max_sched_scan_ssids
networks in the configuration the current logic will attempt to sched
scan a maximal group up to a certain timeout and then switch to
the next group. The timeout also changes with a backoff.

There are a couple of problems with this approach:
	1. The host can suspend while we are sched scanning an arbitrary
	sub group of networks from all the configured networks. The
	device will keep searching only for these networks while suspended
	and other configured networks won't be scanned for.

	2. The current backoff mechanism won't handle well large
	number of networks

Instead of this we switch to a simpler logic whose principles are:
	1. We sched scan only a single group of networks which are determined
	according to those networks with the highest priority. This way
	when the host suspends we'll be scanning the highest priority networks.

	2. If there are more networks than sched scan can handle it will
	fallback to normal scan so we can still scan for the lower priority
	networks.

	3. As there's no need to switch groups there's no need for
	timeout and backoff.

Signed-off-by: Eyal Shapira <eyal@wizery.com>