P2P: ignore neg_req with previously used dialog_token
If for some reason we get a duplicate negotiation request, the
supplicant will generate 2 different responses (different SSIDs)
with the same dialog token. The remote peer will confirm one of them,
but it will probably be the wrong one (the first it received).

Workaround it by ignoring negotiation requests with the last used
dialog_token.

Signed-off-by: Eliad Peller <eliad@wizery.com>