nl80211: reconfigure RX filters whenever a new interface is added (ANDROID)
This is used to reconfigure RX filters to include a new filter whenever
the P2P group interface is added. The new filter matches any unicast
frame sent to a dest mac which matches the P2P group interface mac.
This is important as otherwise we wouldn't allow any data frames through
to this interface.

Signed-off-by: Eyal Shapira <eyal@wizery.com>