dfs: skip ongoing csa if the channel is not available
Under some circumstances, radar event might arrive during
an ongoing channel switch. this can result in a scenario
in which the second ap (still in the first csa) will try
aligning to the channel of the first ap, while its channel
is disabled (because it already updated its channels info).

Add an additional check to make sure the aligned channel
is actually available for use.

Signed-off-by: Eliad Peller <eliad@wizery.com>